concat action and reducers in one fiel(if I got idea right...)
