+++ src/store.js
@@ -1,5 +1,5 @@
 import { createStore, compose, applyMiddleware } from "redux";
-import reducer from "./reducers";
+import reducer from "./reducer";
 import thunk from "redux-thunk";
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
